refactor: ♻️ Extract card field mount logic
